# feat: add `GGML_UNARY_OP_ARGMAX` Metal kernel