--- a/src/ggml-metal/ggml-metal.metal
+++ b/src/ggml-metal/ggml-metal.metal
@@ -1344,6 +1344,23 @@ kernel void kernel_ssm_scan_f32(
     }
 }
 
+kernel void kernel_argmax(
+        device   const void * x,
+        device      int32_t * dst,
+        constant    int64_t & ncols,
+        uint tpitg[[thread_position_in_threadgroup]]) {
+    device const float * x_row = (device const float *) ((device const char *) x + tpitg * ncols * sizeof(float));
+
+    // initialize
+    dst[tpitg] = 0;
+
+    for (int i = 0; i < ncols; i++) {
+        if (x_row[i] > x_row[dst[tpitg]]) {
+            dst[tpitg] = i;
+        }
+    }
+}
+
 kernel void kernel_norm(
         device const  void * src0,
         device       float * dst,