--- a/tests/test-backend-ops.cpp
+++ b/tests/test-backend-ops.cpp
@@ -3440,7 +3440,10 @@ static std::vector<std::unique_ptr<test_case>> make_test_cases_eval() {
     test_cases.emplace_back(new test_conv_transpose_1d({3,2,1,1}, {3,1,2,1}, 1, 0, 1));
     test_cases.emplace_back(new test_conv_transpose_1d({2,1,1,1}, {3,1,1,1}, 1, 0, 1));
 
-    test_cases.emplace_back(new test_argmax());
+    test_cases.emplace_back(new test_argmax(GGML_TYPE_F32, {10, 100 , 1, 1}));
+    test_cases.emplace_back(new test_argmax(GGML_TYPE_F32, {12, 1024, 1, 1}));
+    test_cases.emplace_back(new test_argmax(GGML_TYPE_F32, {3 , 5438, 1, 1}));
+
     test_cases.emplace_back(new test_count_equal());
 
     for (int ne3 : {1, 3}) { // CUDA backward pass only supports ne3 == 1