--- a/tests/test-backend-ops.cpp
+++ b/tests/test-backend-ops.cpp
@@ -3824,6 +3824,10 @@ static std::vector<std::unique_ptr<test_case>> make_test_cases_perf() {
     test_cases.emplace_back(new test_bin_bcast(ggml_add, GGML_TYPE_F32, {4096, 1, 1, 1}, {1,   1, 1, 1}));
     test_cases.emplace_back(new test_bin_bcast(ggml_add, GGML_TYPE_F32, {4096, 1, 1, 1}, {1, 512, 1, 1}));
 
+    test_cases.emplace_back(new test_argmax(GGML_TYPE_F32, {10, 100 , 1, 1}));
+    test_cases.emplace_back(new test_argmax(GGML_TYPE_F32, {12, 1024, 1, 1}));
+    test_cases.emplace_back(new test_argmax(GGML_TYPE_F32, {3 , 5438, 1, 1}));
+
     test_cases.emplace_back(new test_cpy(GGML_TYPE_F32, GGML_TYPE_F16, {512, 3072, 1, 1}));
 
     for (int bs : {1, 512}) {