This PR implements the Metal kernel used for the `GGML_UNARY_OP_ARGMAX` operation.

It is necessary for [Encodec.cpp](https://github.com/PABannier/encodec.cpp) to run on the Metal backend.