@ggerganov 2024-11-16
Thanks, will merge these soon. Just need to sync the https://github.com/ggerganov/llama.cpp/pull/10238 from `llama.cpp` to avoid resolving conflicts manually.