@PABannier 2024-11-19
@ggerganov Thanks for the review. I added your changes:
1. Taking into account the strides to avoid asserting that `src0` is contiguous,
2. Parallelizing over threadgroups instead of threads.

I have a follow-up question regarding the tests/CI: is the `argmax` Metal kernel tested? Looking at the logs I only found one place where `argmax` is mentioned: `19:   ARGMAX(type=f32,ne=[10,100,1,1]): not supported [BLAS]`. Am I missing something?