@PABannier 2024-11-29
@ggerganov I pushed a SIMD implementation of the kernel. All tests are passing :)

I'd like to benchmark both the non-SIMD and SIMD implementations of the kernel. Is there an existing snippet of code that benchmarks the latency and throughput of the kernel as @slaren did in https://github.com/ggerganov/llama.cpp/pull/10441 ?