@slaren 2024-11-29
You can obtain the performance measurements with `test-backend-ops -o ARGMAX perf`. The same tests that I used should already be there, but you can add your own to `make_test_cases_perf`.