@PABannier 2024-11-29
Thanks @slaren ! I run `test-backend-ops -o ARGMAX perf` after adding the 3 test cases.

Here are the results:

With SIMD vectorization:
```
ARGMAX(type=f32,ne=[100,10,1,1]):                   286720 runs -     3.53 us/run -        3 kB/run -    1.07 GB/s
ARGMAX(type=f32,ne=[1024,12,1,1]):                   90112 runs -    11.38 us/run -       48 kB/run -    4.03 GB/s
ARGMAX(type=f32,ne=[5438,3,1,1]):                    65536 runs -    16.11 us/run -       63 kB/run -    3.77 GB/s
```

Without SIMD vectorization:
```
ARGMAX(type=f32,ne=[100,10,1,1]):                    73728 runs -    14.75 us/run -        3 kB/run -    0.26 GB/s
ARGMAX(type=f32,ne=[1024,12,1,1]):                   16384 runs -   120.66 us/run -       48 kB/run -    0.38 GB/s
ARGMAX(type=f32,ne=[5438,3,1,1]):                     8192 runs -   626.23 us/run -       63 kB/run -    0.10 GB/s
```

This is indeed much faster! Thanks for the suggestion @slaren !

This is ok for me on my side if you want to merge :)